+++ src/Zen.Core/Constants/DefaultUrls.cs
@@ -3,5 +3,6 @@ namespace Zen.Core.Constants
     public class DefaultUrls
     {
         public const string IFCONFIG_URL = "http://ifconfig.me/all.json";
+        public const string GITIGNORE_IO_URL = "https://www.gitignore.io";
     }
 }